feat: Adicionando Serializer global
