+++ src/main/kotlin/application/domain/models/Income.kt
@@ -1,13 +1,8 @@
 package application.domain.models
 
 import kotlinx.serialization.Contextual
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
+import starter.serializer.BigDecimalSerializer
 import java.math.BigDecimal
 
 @Serializable
@@ -17,7 +12,7 @@ data class Income(
     val value: BigDecimal
 ): ValueObject {
     init {
-        require(value >= MIN_VALUE) { "Amount less than $MIN_VALUE" }
+        require(value >= MIN_VALUE) { "Income less than $MIN_VALUE" }
     }
 
     companion object {
@@ -27,15 +22,3 @@ data class Income(
     }
 }
 
-object BigDecimalSerializer : KSerializer<BigDecimal> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("BigDecimal", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: BigDecimal) {
-        encoder.encodeString(value.toString())
-    }
-
-    override fun deserialize(decoder: Decoder): BigDecimal {
-        return BigDecimal(decoder.decodeString())
-    }
-}
