+++ src/main/kotlin/driven/database/LoanDto.kt
@@ -1,21 +1,20 @@
 package driven.database
 
 import kotlinx.serialization.Contextual
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
+import starter.serializer.LocalDateTimeSerializer
 import java.time.LocalDateTime
 
 @Serializable
 data class LoanDto (
     val loanId: String,
-    val customer: String,
+    val customer: String? = null,
     val status: String,
     val version: Int,
+    val proposals: String? = null,
+    val type: String? = null,
+//    val amount:  BigDecimal? = null
+//    val tax: BigDecimal? = null
     @Serializable(with = LocalDateTimeSerializer::class)
     @Contextual
     val createdAt: LocalDateTime? = null,
@@ -24,15 +23,3 @@ data class LoanDto (
     val updatedAt: LocalDateTime? = null
 )
 
-object LocalDateTimeSerializer : KSerializer<LocalDateTime> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("LocalDateTime", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: LocalDateTime) {
-        encoder.encodeString(value.toString())
-    }
-
-    override fun deserialize(decoder: Decoder): LocalDateTime {
-        return LocalDateTime.parse(decoder.decodeString())
-    }
-}
